requirements.txt
# Indian Apple Intelligence System - Dependencies
# ============================================
# All free, open-source packages for offline operation

# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Web Interface
gradio>=3.50.0

# Data Processing
numpy<2.0.0,>=1.24.0
pandas>=2.0.0

# Visualization (required for training)
matplotlib>=3.7.0
seaborn>=0.12.0

# System Utilities
tqdm>=4.65.0

# Note: All packages are free and open-source
# No premium APIs or cloud services required
# System runs 100% offline after installation
